requirements.txt
flask==2.1.2
pandas==1.3.5
numpy==1.22.3
scipy==1.8.1
surprise==0.1
mysql-connector-python==8.0.27
